requirements-dev.txt
black==21.6b0
isort==5.8.0
pre-commit==2.13.0
ipdb==0.13.9
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-cov==2.12.1
